pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "attrib-regression"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "numpy==2.4.1",
    "pandas==2.3.3",
    "scipy==1.17.0",
    "scikit-learn==1.8.0",
    "statsmodels==0.14.6",
    "matplotlib==3.10.8",
    "seaborn==0.13.2",
    "PyYAML==6.0.3",
    "joblib==1.5.3",
    "tqdm==4.67.1",
    "jupyter==1.1.1",
    "ipykernel==7.1.0",
    "google-cloud-bigquery==3.40.0",
    "google-cloud-bigquery-storage==2.36.0",
    "openpyxl==3.1.5",
    "pyarrow==23.0.0",
    "db-dtypes==1.5.0",
    "pandas-stubs~=2.3.3",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "nbqa",
    "pytest",
]

[project.scripts]
rba-pipeline = "attrib_regression.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"notebooks/**/*.ipynb" = ["E402", "F401", "F841"]

[tool.nbqa.addopts]
ruff = ["--ignore=E402,F401,F841"]

[tool.pytest.ini_options]
testpaths = ["tests"]
